feat: check if cat's age did not violate rules

+++ internal/models/cat.go
@@ -12,7 +12,7 @@ type Cat struct {
 	Name        string    `json:"name" db:"name"`
 	Race        string    `json:"race" db:"race"`
 	Sex         string    `json:"sex" db:"sex"`
-	AgeInMonth  uint16    `json:"ageInMonth" db:"age_in_month"`
+	AgeInMonth  int32     `json:"ageInMonth" db:"age_in_month"`
 	Description string    `json:"description" db:"description"`
 	ImageUrls   []string  `json:"imageUrls" db:"image_urls"`
 }
